refactor: 태그 관리기능 리팩토링
