--- src/main/java/homeTry/adminTag/productTag/dto/response/ProductTagResponse.java
@@ -1,10 +1,0 @@
-package homeTry.adminTag.productTag.dto.response;
-
-import java.util.List;
-
-import homeTry.adminTag.productTag.dto.ProductTagDto;
-
-public record ProductTagResponse(
-    List<ProductTagDto> productTags
-) {
-}
