--- src/main/java/homeTry/adminTag/productTag/exception/BadRequestException/ProductTagNotFoundException.java
@@ -1,11 +1,0 @@
-package homeTry.adminTag.productTag.exception.BadRequestException;
-
-import homeTry.adminTag.productTag.exception.ProductTagErrorType;
-import homeTry.common.exception.BadRequestException;
-
-public class ProductTagNotFoundException extends BadRequestException{
-    
-    public ProductTagNotFoundException() {
-        super(ProductTagErrorType.PRODUCT_TAG_NOT_FOUND_EXCEPTION);
-    }
-}
