--- src/main/java/homeTry/adminTag/teamTag/dto/response/TeamTagResponse.java
@@ -1,9 +1,0 @@
-package homeTry.adminTag.teamTag.dto.response;
-
-import java.util.List;
-
-import homeTry.adminTag.teamTag.dto.TeamTagDTO;
-
-public record TeamTagResponse(
-    List<TeamTagDTO> teamTags) {
-}
