--- src/main/java/homeTry/adminTag/teamTag/exception/BadRequestException/TeamTagNotFoundException.java
@@ -1,12 +1,0 @@
-package homeTry.adminTag.teamTag.exception.BadRequestException;
-
-import homeTry.adminTag.teamTag.exception.TeamTagErrorType;
-import homeTry.common.exception.BadRequestException;
-
-public class TeamTagNotFoundException extends BadRequestException {
-    
-    public TeamTagNotFoundException() {
-        super(TeamTagErrorType.TEAM_TAG_NOT_FOUND_EXCEPTION);
-    }
-    
-}
