--- src/main/java/homeTry/adminTag/teamTag/repository/TeamTagRepository.java
@@ -1,9 +1,0 @@
-package homeTry.adminTag.teamTag.repository;
-
-import homeTry.adminTag.teamTag.model.entity.TeamTag;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface TeamTagRepository extends JpaRepository<TeamTag, Long> {
-}
