+++ src/main/java/homeTry/tag/productTag/dto/response/ProductTagResponse.java
@@ -1,0 +1,10 @@
+package homeTry.tag.productTag.dto.response;
+
+import java.util.List;
+
+import homeTry.tag.productTag.dto.ProductTagDto;
+
+public record ProductTagResponse(
+    List<ProductTagDto> productTags
+) {
+}
