+++ src/main/java/homeTry/tag/productTag/repository/ProductTagRepository.java
@@ -1,0 +1,9 @@
+package homeTry.tag.productTag.repository;
+
+import homeTry.tag.productTag.model.entity.ProductTag;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ProductTagRepository extends JpaRepository<ProductTag, Long>{
+}
