--- src/main/java/homeTry/tag/repository/TagRepository.java
@@ -1,8 +1,0 @@
-package homeTry.tag.repository;
-
-import homeTry.tag.model.entity.Tag;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TagRepository extends JpaRepository<Tag, Long> {
-
-}
