--- src/main/java/homeTry/tag/repository/TeamTagRepository.java
@@ -1,7 +1,0 @@
-package homeTry.tag.repository;
-
-import homeTry.tag.model.entity.TeamTag;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeamTagRepository extends JpaRepository<TeamTag, Long> {
-}
