+++ src/main/java/homeTry/tag/teamTag/dto/response/TeamTagResponse.java
@@ -1,0 +1,10 @@
+package homeTry.tag.teamTag.dto.response;
+
+import homeTry.tag.teamTag.dto.TeamTagDTO;
+
+import java.util.List;
+
+
+public record TeamTagResponse(
+    List<TeamTagDTO> teamTags) {
+}
