+++ src/main/java/homeTry/tag/teamTag/repository/TeamTagRepository.java
@@ -1,0 +1,7 @@
+package homeTry.tag.teamTag.repository;
+
+import homeTry.tag.teamTag.model.entity.TeamTag;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface TeamTagRepository extends JpaRepository<TeamTag, Long> {
+}
