+++ src/main/java/homeTry/team/service/TeamService.java
@@ -5,9 +5,9 @@ import homeTry.exerciseList.service.ExerciseTimeService;
 import homeTry.member.dto.MemberDTO;
 import homeTry.member.model.entity.Member;
 import homeTry.member.service.MemberService;
-import homeTry.tag.dto.TeamTagDTO;
-import homeTry.tag.model.entity.TeamTag;
-import homeTry.tag.service.TeamTagService;
+import homeTry.tag.teamTag.dto.TeamTagDTO;
+import homeTry.tag.teamTag.model.entity.TeamTag;
+import homeTry.tag.teamTag.service.TeamTagService;
 import homeTry.team.dto.RankingDTO;
 import homeTry.team.dto.request.CheckingPasswordRequest;
 import homeTry.team.dto.request.TeamCreateRequest;
